Update patch set 2
Patch Set 2:

I like the navbar because I frequently change between the S2 and a tablet. Having the same navigation elements is very convenient. And I also like to pin an app, which cannot be released without the navbar (on our S2). Having a directly accessible app switch is great (without sacrifying the menu key if you did it without navbar).
And what is the disadvantage of being able to select according to one's own taste instead of being urged to the learn what is "correct"?
With my proposed solution the capacitive keys are really disabled, so there should be no risk of confusing people.

Patch-set: 2